Add handling for absolute asset paths
Previously the task was only working for relative paths. If you tried to
fingerprint a path like /css/main.css, the boot file wouldn't be matched
and the task would fail trying to calculate a SHA of nil.

Now the absolute path prefixes are stripped out for file matching and
placed back in for the final output.

I tested that this worked using the relative path "css/main.css" as well
as the absolute path "/css/main.css"